fix: Karten-Produktempfehlung diversifizieren + Backtest-Tab wiederherstellen
1. region_tooltip_service: _infer_condition() differenziert jetzt nach
   peix_band + trend + zweitstärkstem Driver statt nur Versorgungslage
   (war 100% bei allen Regionen → immer GeloMyrtol forte)
2. Backtest als 4. Navigation-Tab wiederhergestellt (Markt-Check,
   CSV-Upload, Recharts-Chart, Verlaufs-Tabelle)
3. backtester.py: decimal.Decimal/float Typenkonflikt gefixt

+++ backend/app/services/ml/backtester.py
@@ -178,7 +178,7 @@ class BacktestService:
         ).scalar() or 0
 
         if previous > 0:
-            slope = (recent - previous) / previous
+            slope = float((recent - previous) / previous)
         else:
             slope = 0.0
